+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - add SelectNodePlugin that can interrupt the Select method
 - support TcpCopy by TeePlugin
 - Change two methods' name of client plugin!
+- Broken API: add error parameter in `PreWriteResponse`(#486)
 
 ## 5.0